misc/HtmlTestSuite/results/test-0059.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head><meta charset="utf-8"/>
<title>Event Timing: eventCounts.</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/testdriver.js"/>
<script src="/resources/testdriver-vendor.js"/>
</head><body><div id="div">Click me</div>
<button id="button">Click me</button>
<script>
  promise_test( t =&gt; {
    assert_implements(window.EventCounts, "Event Counts isn't supported");
    function testClicks(expectedCount, resolve) {
      const clickCount = performance.eventCounts.get('click');
      if (clickCount &lt; expectedCount) {
        t.step_timeout(function() {
          testClicks(expectedCount, resolve);
        }, 5);
        return;
      }
      assert_equals(clickCount, expectedCount,'Incorrect click count.');
      assert_equals(performance.eventCounts.get('mousedown'), expectedCount, 'Incorrect mousedown count');
      assert_equals(performance.eventCounts.get('mouseup'), expectedCount, 'Incorrect mouseup count.');
      resolve();
    }
    function promiseClicks(expectedCount) {
      return new Promise(resolve =&gt; {
        testClicks(expectedCount, resolve)
      });
    }

    return test_driver.click(document.getElementById('div')).then(() =&gt; {
      return promiseClicks(1);
    }).then(() =&gt; {
      return test_driver.click(document.getElementById('button'));
    }).then(() =&gt; {
      return promiseClicks(2);
    }).then(() =&gt; {
      return test_driver.click(document.getElementById('div'));
    }).then(() =&gt; {
      return promiseClicks(3);
    });
  })
</script>
</body></html>
